fixed  duplicate prop naming
